threadlocal util
The custom celery task class will grab the transaction_id
from the threadlocaldict which is put there inside middleware.

One thing in particular that this means, at this point, is that
the transaction_id that we are to now begin using is representative
of an event that must begin at the API layer. IOW, we are not
generating transaction ids anywhere else in the codebase.