config: handle '--' being first argument
The change in 35e002d introduced special handling for '--' being the
first argument. In that case,

  return argv[1:], []

was returned. I believe the correct return value would be

  return [], argv[1:]

since the non-options are the second return value. When the return
values are swapped, the special-casing can be removed since the
generic case also covers the case where '--' is the first argument.