fix(test_common_git): add a repo `commit` fix
A test which was fetching from a "remote" repo and checking out a commit
from it was failing, but only with an older version of git (1.7.x.x).
After doing some research, it seems this is a problem with `git
rev-parse` and was fixed in version >= 1.9. The problem is with a
missing .git/refs/heads/master file in a freshly-inited repo.

Adding a `repo.commit()` works around this issue.

I admit that I'm too lazy right now to figure out exactly which version
of git fixes the issue. If someone can do the research, a follow-up
patch would be useful. =)