feat(metaconfig): for new config sources
meta config options allow an entire configuration to be
specified by a single source. the only existing native
capability was the use of ini files, but there was no
way to specify which ini file, much less specify it the
same way any other Option could be specified (cli, env, keyring).

this connects a metaconfig instance that can be used to provision
an entire (or part of) a Config. the first meta option to be
implemented was `--ini` since Config natively supports parsing ini
files.

See the MetaConfig docstring for more details.