add copy and repr for Option
Adding __repr__ makes it easier to figure out which
Option(s) you're looking at when debugging or showing
the output of the Options list.

Adding __copy__ allows us to make a copy of the options
list so that we can modify it for other purposes. The
initial use case was to extract the defaults using
argparse and "temporarily" popping the 'required' keyword
argument off each Option.

To make a copy of the full list of options, without having
to worry about mutating them in a destructive way,

    opts = [copy.copy(o) for o in self._options]