use custom GitWarning for better control
Tests were failing on travis because travis uses git 1.8.5.6
When git.py was imported, the warning was emitted. Then,
when the tests ran, the warning had already been issued, so
by default python filtered it out and did not display it, causing
the test to fail since we assert that the warning was emitted.

Instead of modifying the filtering behavior of RuntimeWarning,
it seemed more reasonable to modify the filtering behavior
of only our own warnings, so GitWarning (a RuntimeWarning)
was created, and upon importing simpl the GitWarning filter
is changed to 'always' show. This can easily be overridden
by a subsequent filter or by calling the GitWarning's disable()
method.