chore(style): update pylint settings
The main checks we want pylint to do are C, E, and
F (Convention, Error, and Fatal, respectively).
There are a few exceptions. The R (Refactor)
warnings are ostensibly useful, but they're mostly
just "too-many-X" warnings, which I find to be a
bit silly and arbitrary.

By playing around with the config and tweaking a
couple of things, I came up with this small patch
that gives us a perfect score, but keeps many of
the critical checks in place.