feat(rest): add error formatter
Bottle returns error information by default as an
HTML page. But for writing APIs (which is what
simpl is used for mostly) we need to return
API-parseable code. This formatter ensures errors
are returned in the format requested by the client
and using a known format (dict with predictable
keys and schema).