fix(rest): 416 on unexpected content-types
Because this explicit error isn't raised, I found subtle bugs in an
application that uses this. If the client forgot to specify the
content-type in the request header, the server would process the request
as if everything was fine, except that it would ignore the content. The
was leading to incorrect behavior in the server application.