feat: error handling/formatting middleware
Supports json and yaml response content types.
Supports handling bottle's httperror and simpl's
httperrors with the same logic.

The suggested use is to run a bottle app with this
middleware and set catchall to False. If catchall
is *not* set to False, this middleware will rarely
get used for catching generic exceptions and
SimplHTTPErrors unless those exceptions are raised
in other middleware and thus cannot be caught by
bottle's catchall.

https://github.com/bottlepy/bottle/blob/0.12.8/bottle.py#L874
https://github.com/bottlepy/bottle/blob/0.12.8/bottle.py#L933