consolidate --debug and --quiet
In addition to configuring log format and log level, the use
of --debug will also set the bottle server to debug mode, which
will return tracebacks in the response body if an unexpected error
occurs.

If --quiet is used, in addition to configuring a minimal log format
and WARNING log level, the routes and other startup text is not printed
to stdout.